fixed backup & project, appointment data

+++ php/backup.php
@@ -4,16 +4,25 @@ ini_set( 'display_errors', 1 );
 ini_set( 'display_startup_errors', 1 );
 error_reporting( E_ALL );
 
-$folder_to_backup = '../userdata';
+$folder_to_backup = '../';
 $backup_folder    = 'backups';
 //  array($backup_folder, 'privat', 'config');
 $not_to_backup    = [$backup_folder];
-$backup_file_name = $backup_folder.'/'.date( "Y_m_d" ).'.zip';
+$backup_file_name = $backup_folder.'/'.date( "Y_m_d_H_i" ).'.zip';
 
 // create the backup_folder if it not exist
 if ( !file_exists( $backup_folder ) ) {mkdir( $backup_folder, 0777, true );}
 
 // Backup function
+/**
+ * This function creates a zip archive of a folder and all its contents
+ *
+ * @param folder_to_backup The folder you want to backup.
+ * @param backup_file_name The name of the backup file.
+ * @param not_to_backup an array of files and folders that you don't want to backup.
+ *
+ * @return A boolean value.
+ */
 function run_backup( $folder_to_backup, $backup_file_name, $not_to_backup ) {
     // Remove existing archive
     if ( file_exists( $backup_file_name ) ) {unlink( $backup_file_name );}
@@ -48,26 +57,30 @@ function run_backup( $folder_to_backup, $backup_file_name, $not_to_backup ) {
 
 // get size of the lastest backup file
 $files = scandir( $backup_folder, SCANDIR_SORT_DESCENDING );
-// print_r( $files );
 
-foreach ( $files as $file ) {
-    if ( date( "Y_m_d" ).'.zip' === $file ) {
-        // backup today already done
-        exit;
+if ( $files ) {
+    // check for backup today
+    foreach ( $files as $file ) {
+        if ( date( "Y_m_d_H_i" ).'.zip' === $file ) {
+            // backup today already done
+            exit;
+        }
     }
+    $last_file_size = filesize( $backup_folder.'/'.$files[0] );
 }
-$last_file_size = filesize( $backup_folder.'/'.$files[0] );
 
 // run di dance
 run_backup( $folder_to_backup, $backup_file_name, $not_to_backup );
+if ( $files ) {
 
 // get size of this backupfile
-$files          = scandir( $backup_folder, SCANDIR_SORT_DESCENDING );
-$this_file_name = $files[0];
-$this_file_size = filesize( $backup_folder.'/'.$files[0] );
+    $files          = scandir( $backup_folder, SCANDIR_SORT_DESCENDING );
+    $this_file_name = $files[0];
+    $this_file_size = filesize( $backup_folder.'/'.$files[0] );
 
-// delete this backup, if the same size like last backup
-if ( $last_file_size == $this_file_size ) {
-    unlink( $backup_folder.'/'.$files[0] );
+    // delete this backup, if the same size like last backup
+    if ( $last_file_size == $this_file_size ) {
+        unlink( $backup_folder.'/'.$files[0] );
+    }
 }
 // echo 'no changes';
